pyproject.toml
[tool.poetry]
name = "rakuten-fr-multimodel"
version = "0.1.0"
description = "Model to predict product categories based on images and text (Title+description)"
authors = ["Rudolf Braun <48672663+BraunRudolf@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
torch = "^2.4.0"
pandas = "^2.2.2"
nltk = "^3.8.1"
beautifulsoup4 = "^4.12.3"
scikit-learn = "^1.5.1"
spacy = "^3.7.5"
mlflow = "^2.15.0"
setuptools = "^72.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
logging = "^0.4.9.6"
tqdm = "^4.66.4"
torchvision = "^0.19.0"


[tool.poetry.scripts]
rakuten = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
